+++ NOTES.md
@@ -6,27 +6,33 @@ Corpses should become piles? Not sure yet. But this fun would be nice to have...
 
 # Todo
 
-- [ ] Multiple dungeon floors
 - [ ] Improve AI
 - [ ] Weapons should have variable damage
 - [ ] Chance to hit
 - [ ] minimum damage so a hit can't do 0 damage (should always be some sort of a threat even if it's minimal)
 - [ ] food & drink
 - [ ] shops
-- [ ] sort items in container UI
 - [ ] some variablility to init rendering of dungeon - not static numbers of entities as we currently have
 - [ ] character tab for player stats
+- [ ] convert wallet to currency on death for monsters
+
+# Bugs
+
+- [ ] Inventory does not carry through to next level - different player instance per floor! oops! Duh. I'm reinitializing the entire game state. So a brand new player entity. There is literally a different player per floor...
 - [ ] items on floor have Wear and Wield actions but you cant actually do them until you pick them up
+- [ ] items in container UI are not sorted
 
 # Doing
 
+- [ ] Only allow ascend and descend when on stairs
+
+# Done
+
+- [x] Multiple dungeon floors
 - [x] money!
 - [x] add money to UI somehere
 - [x] increment money on currency pickup
 - [x] use scribe coinpurse for money counting
-- [ ] pick up money from chests
-- [ ] convert wallet to currency on death for monsters
-
-# Done
+- [x] pick up money from chests
 
 Gold is an entity (item) that happens to not go in your inventory but incremnet your gold stat - like an instant potion...
