+++ src/ECS/systems/render.system.js
@@ -374,20 +374,6 @@ const renderContainer = () => {
 
   drawItemList();
   drawSelectedItemDetails();
-
-  // look for any that have an inventory. If they do:
-  // need to be able to remove item from entity inventory (so build some data structire that will faciliate that)
-  // items: {
-  //   floor: []
-  //   entityId: [] // id of entity whose inventory item belongs to
-  //   ...
-  // }
-  // lists all items
-  // store current selected item
-  // on get
-  // remove from floor or inventory
-  // add to player inventory
-  // all this needs to somehow be able to work for monsters npcs to do this without a UI...
 };
 
 const renderInventory = () => {
